poprawki potrzebne do zgodnosci z main

+++ src/main/java/sggw/wzim/czasnawypad/repository/AttractionRatingRepository.java
@@ -6,12 +6,13 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import sggw.wzim.czasnawypad.model.AttractionRating;
+import sggw.wzim.czasnawypad.db.entity.AttractionRating;
+import sggw.wzim.czasnawypad.db.entity.User;
 
 @Repository
-public interface AttractionRatingRepository extends JpaRepository<AttractionRating, Long> {
+public interface AttractionRatingRepository extends JpaRepository<AttractionRating, Integer> {
     List<AttractionRating> findByUser(User user);
-    List<AttractionRating> findByAttractionId(Long attractionId);
+    List<AttractionRating> findByAttractionId(Integer attractionId);
     
     @Query("SELECT r.attraction.id AS attractionId, AVG(r.rating) AS averageRating " +
            "FROM AttractionRating r " +
